rename module_struct file to container_strust
